feat: use once_cell to ensure state is only initialized once
